fix: Cast GetFabricIndex() return value to jint in AndroidLogDownloadFromNode
The change adds an explicit static cast to convert the FabricIndex
return value to jint type when assigning to jFabricIndex variable,
maintaining type safety in the JNI interface layer.

+++ src/controller/java/AndroidLogDownloadFromNode.cpp
@@ -210,7 +210,7 @@ void AndroidLogDownloadFromNode::FinishLogDownloadFromNode(void * context, CHIP_
     JniLocalReferenceScope scope(env);
 
     jobject jCallback   = self->mJavaCallback.ObjectRef();
-    jint jFabricIndex   = self->mController->GetFabricIndex();
+    jint jFabricIndex   = static_cast<jint>(self->mController->GetFabricIndex());
     jlong jremoteNodeId = static_cast<jlong>(self->mRemoteNodeId);
 
     VerifyOrExit(env != nullptr, ChipLogError(Controller, "Could not get JNIEnv for current thread"));
